implement wayfire network keyboard
Network keyboard consists of two programs - server and client.

The server sits on one computer and grabs all keyboard input, sending it
over to one client. The client receives the keyboard events, and
emulates a keyboard device via the virtual-keyboard-v1 protocol.

Example usage:

On the server, say 192.168.0.100 and port 12345:

wf-nk-server 12345

On the client, in the same network:

wf-nk-client 192.168.0.100 12345