Add Follow-Focus plugin
When this plugin is enabled, the input focus will follow the mouse
pointer. Hence, when moving the mouse and typing afterwards, the typed
keys will be send to the view that is currently under the cursor.
However, this will not change the stacking order of the views.

This behavior is similar to i3/sway or other X-based DEs.