keycolor: Store program data on core
Storing the program as a global variable causes wayfire to crash
when disabling an output and in general is bad practice. Here we
store the program as custom data which is attatched to core, since
we only need a single instance of the object. This let's us manage
it so that the lifetime begins with the first plugin instance and
ends with the last.