follow-focus: Make threshold more consistent
The plugin was counting from the point where it last focused a view,
and not resetting the point when entering a different view. This caused
the threshold to sometimes work and other times not. Fix this by tracking
the last view and resetting the point from which we are measuring, when
moving to a new view.