follow-focus: Use output objects instead of core to focus views
This makes it so we can properly optimize the case where we have already
focused the view the cursor is moving over. It sets the internal active_view
variable for the output so that output->get_active_view() returns the view
with keyboard focus, instead of the topmost view. We compare this with the
cursor focus view and if they match, do nothing.