Focus change (#219)
This plugin allows the user to cycle through views using keybindings. This is similar to features seen in tiling window managers except it doesn't rely on tiling functionality.

The plugin provides a fair few options adjustable to the user so they can tweak it to their liking. The code hasn't successfully been tested on multiple outputs, and potentially the code related to the 'cross-output' settings isn't working properly.