background-view: Drop (#247)
This plugin is replaced by pin-view plugin, which is ipc based. To use it,
'pip3 install wayfire' and find the scripts in this repo under the ipc-scripts
directory. [un]pin-view.py can handle multiple views, and different views on
different workspaces. It is recommended to use in conjuntion with the ghost
plugin so the pinned views do not recieve input.