Add extra-animations plugin (#265)
The extra-animations plugin adds four new animations - blinds, helix, shatter and vortex.
It requires animate and extra-animations plugins to be enabled. Currently, the new animations
will not show up in wcm animate plugin, the config file must be edited manually. However,
the options for the animations can be adjusted in wcm extra animations plugin. The
animations are valid for open and close animation options.