annotate: fix unloading (#281)
* annotate: fix unloading

When annotate is loaded on non-GLES renderer, it will not initialize the
overlay array. So when we iterate at the end, it will crash as the array
is empty. Instead, we can iterate over all the elements in the array.
This way, it works also for empty arrays :)

In addition I also removed a few of the .reset() calls. They are not
necessary, as the memory will be freed automatically by the smart
pointers (shared_ptrs, unique_ptrs).

* fixup! annotate: fix unloading